requirements.txt
streamlit>=1.31.0
langchain>=0.3.0
langchain-classic>=1.0.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
faiss-cpu>=1.8.0
pypdf>=4.2.0
python-dotenv>=1.0.1
